// apps/web/src/api/generated/schemas/appointmentCreate.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { AppointmentCreateClinicianId } from './appointmentCreateClinicianId';
import type { AppointmentCreateBranchId } from './appointmentCreateBranchId';
import type { AppointmentType } from './appointmentType';
import type { AppointmentStatus } from './appointmentStatus';
import type { AppointmentCreateNotes } from './appointmentCreateNotes';

export interface AppointmentCreate {
  partyId: string;
  clinicianId?: AppointmentCreateClinicianId;
  branchId?: AppointmentCreateBranchId;
  date: string;
  /** @pattern ^\d{2}:\d{2}$ */
  time: string;
  duration?: number;
  appointmentType?: AppointmentType;
  status?: AppointmentStatus;
  notes?: AppointmentCreateNotes;
}