// apps/web/src/api/generated/schemas/responseEnvelopeListAppointmentRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseEnvelopeListAppointmentReadData } from './responseEnvelopeListAppointmentReadData';
import type { ResponseEnvelopeListAppointmentReadMessage } from './responseEnvelopeListAppointmentReadMessage';
import type { ResponseEnvelopeListAppointmentReadError } from './responseEnvelopeListAppointmentReadError';
import type { ResponseEnvelopeListAppointmentReadMeta } from './responseEnvelopeListAppointmentReadMeta';
import type { ResponseEnvelopeListAppointmentReadRequestId } from './responseEnvelopeListAppointmentReadRequestId';

export interface ResponseEnvelopeListAppointmentRead {
  success?: boolean;
  data?: ResponseEnvelopeListAppointmentReadData;
  message?: ResponseEnvelopeListAppointmentReadMessage;
  error?: ResponseEnvelopeListAppointmentReadError;
  meta?: ResponseEnvelopeListAppointmentReadMeta;
  requestId?: ResponseEnvelopeListAppointmentReadRequestId;
  timestamp?: string;
}